requirements.txt
requests
pandas
sqlalchemy
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
schedule
psycopg2-binary 